requirements.txt
# ─── Core ──────────────────────────────────────────────────────────────────────
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.28.0
datasets>=2.18.0
tokenizers>=0.15.0

# ─── Embedding & Retrieval ─────────────────────────────────────────────────────
sentence-transformers>=3.0.0
faiss-cpu>=1.7.4          # GPU varsa: faiss-gpu
rank_bm25>=0.2.2
chromadb>=0.5.0

# ─── Fine-tuning (QLoRA) ───────────────────────────────────────────────────────
peft>=0.10.0
bitsandbytes>=0.43.0
trl>=0.8.6

# ─── Evaluation ────────────────────────────────────────────────────────────────
ragas>=0.1.14
rouge-score>=0.1.2
sacrebleu>=2.4.0
evaluate>=0.4.1
bert-score>=0.3.13

# ─── Data & Utils ─────────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.0
huggingface-hub>=0.22.0
kaggle>=1.6.0

# ─── Experiment Tracking ──────────────────────────────────────────────────────
wandb>=0.17.0

# ─── Demo ─────────────────────────────────────────────────────────────────────
gradio>=4.36.0

# ─── Notebook ─────────────────────────────────────────────────────────────────
jupyter>=1.0.0
ipykernel>=6.29.0
matplotlib>=3.8.0
seaborn>=0.13.0
